postgres-queries-vs-functions-performance/benchmarks/requirements.txt
psycopg2-binary==2.9.5
tqdm==4.64.1
